+++ src/main/java/org/graphysica/espace2d/forme/Dessinable.java
@@ -14,10 +14,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graphysica.espace2d;
+package org.graphysica.espace2d.forme;
 
 import com.sun.istack.internal.NotNull;
 import javafx.scene.canvas.Canvas;
+import org.graphysica.espace2d.Repere;
 
 /**
  * Les classes implémentant cette interface peuvent être dessinées dans un
